+++ test/python/test_MDLCSR_Broadcast.cpp
@@ -98,14 +98,14 @@ void runTest(const char* elFile, std::uint64_t numVertices) {
       fmapVoid(masterBitSet, fill, true);
       barrier.done();
     };
-    for (std::int16_t nodeId = 0; nodeId < dims.node.id; nodeId++) {
+    for (std::int64_t nodeId = 0; nodeId < dims.node.id; nodeId++) {
       PANDO_CHECK(
           pando::executeOn(pando::Place{pando::NodeIndex{nodeId}, pando::anyPod, pando::anyCore},
                            func, barrier, graph.getMasterBitSets()));
     }
     PANDO_CHECK(barrier.wait());
 
-    for (std::int16_t nodeId = 0; nodeId < dims.node.id; nodeId++) {
+    for (std::int64_t nodeId = 0; nodeId < dims.node.id; nodeId++) {
       pando::GlobalRef<Graph::LocalVertexRange> masterRange = graph.getMasterRange(nodeId);
       for (Graph::VertexTopologyID masterTopologyID = *lift(masterRange, begin);
            masterTopologyID < *lift(masterRange, end); masterTopologyID++) {
@@ -116,7 +116,7 @@ void runTest(const char* elFile, std::uint64_t numVertices) {
 
     graph.broadcast();
 
-    for (std::int16_t nodeId = 0; nodeId < dims.node.id; nodeId++) {
+    for (std::int64_t nodeId = 0; nodeId < dims.node.id; nodeId++) {
       pando::GlobalRef<pando::Array<bool>> mirrorBitSet = graph.getMirrorBitSet(nodeId);
       pando::GlobalRef<pando::Array<Graph::MirrorToMasterMap>> localMirrorToRemoteMasterOrderedMap =
           graph.getLocalMirrorToRemoteMasterOrderedMap(nodeId);
